feat: ✨ move config to its own file
