+++ apps/api/src/routes/v1/rest/websoc/+endpoint.ts
@@ -1,16 +1,7 @@
-import { ApiPropsOverride } from "@bronya.js/api-construct";
 import { PrismaClient } from "@libs/db";
 import { createErrorResult, createOKResult } from "@libs/lambda";
 import type { WebsocAPIResponse } from "@libs/uc-irvine-api/websoc";
 import { combineAndNormalizeResponses, notNull, sortResponse } from "@libs/websoc-utils";
-import {
-  Effect,
-  ManagedPolicy,
-  PolicyDocument,
-  PolicyStatement,
-  Role,
-  ServicePrincipal,
-} from "aws-cdk-lib/aws-iam";
 import type { APIGatewayProxyHandler } from "aws-lambda";
 import { ZodError } from "zod";
 
@@ -142,31 +133,4 @@ export const GET: APIGatewayProxyHandler = async (event, context) => {
     }
     return createErrorResult(400, error, requestId);
   }
-};
-
-// TODO: move this into a separate file.
-export const overrides: ApiPropsOverride = {
-  constructs: {
-    functionProps: (scope, id) => ({
-      role: new Role(scope, `${id}-v1-rest-websoc-role`, {
-        assumedBy: new ServicePrincipal("lambda.amazonaws.com"),
-        managedPolicies: [
-          ManagedPolicy.fromAwsManagedPolicyName("service-role/AWSLambdaBasicExecutionRole"),
-        ],
-        inlinePolicies: {
-          lambdaInvokePolicy: new PolicyDocument({
-            statements: [
-              new PolicyStatement({
-                effect: Effect.ALLOW,
-                resources: [
-                  `arn:aws:lambda:${process.env["AWS_REGION"]}:${process.env["ACCOUNT_ID"]}:function:peterportal-api-next-services-prod-websoc-proxy-function`,
-                ],
-                actions: ["lambda:InvokeFunction"],
-              }),
-            ],
-          }),
-        },
-      }),
-    }),
-  },
 };
